Found the error for the autograder and removed it
The makefile lines about todays_date needed to be removed as those files didn't exist in the lab and were only in the practice during recitation.  since those files didn't exist, it was failing on those compiling lines, and stopped trying to compile the lines after, which were in the lab and very much required.  
Only change was to the Makefile, the leftover code from the practice is still in the code documents, just unused.